Conflicts:
	.gitignore
	Application/VixenApplication/VixenApplication.csproj
	Common/BaseSequence/BaseSequence.csproj
	Common/BasicInputManagement/BasicInputManagement.csproj
	Common/Controls/Controls.csproj
	Common/ScriptSequence/ScriptSequence.csproj
	Modules/App/ColorGradients/ColorGradients.csproj
	Modules/App/Curves/Curves.csproj
	Modules/App/DisplayPreviewModule/DisplayPreview.csproj
	Modules/App/InputEffectRouter/InputEffectRouter.csproj
	Modules/App/InstrumentationPanel/InstrumentationPanel.csproj
	Modules/App/Scheduler/Scheduler.csproj
	Modules/Editor/ScriptEditor/ScriptEditor.csproj
	Modules/Editor/TimedSequenceEditor/TimedSequenceEditor.csproj
	Modules/Editor/TimedSequenceEditor/TimedSequenceEditorForm.Designer.cs
	Modules/Editor/TimedSequenceEditor/TimedSequenceEditorForm.cs
	Modules/Editor/TimedSequenceEditor/TimedSequenceElement.cs
	Modules/Effect/Candle/Candle.csproj
	Modules/Effect/Chase/Chase.csproj
	Modules/Effect/ColorWithIntent/ColorWithIntent.csproj
	Modules/Effect/ImageGrid/ImageGrid.csproj
	Modules/Effect/Pulse/Pulse.csproj
	Modules/Effect/RampWithIntent/RampWithIntent.csproj
	Modules/Effect/SetLevel/SetLevel.csproj
	Modules/Effect/SetPosition/SetPosition/SetPosition.csproj
	Modules/Effect/Spin/Spin.csproj
	Modules/Effect/Twinkle/Twinkle.csproj
	Modules/EffectEditor/ChaseEffectEditor/ChaseEffectEditor.csproj
	Modules/EffectEditor/ColorGradientTypeEditor/ColorGradientTypeEditor.csproj
	Modules/EffectEditor/ColorTypeEditor/ColorTypeEditor.csproj
	Modules/EffectEditor/CurveTypeEditor/CurveTypeEditor.csproj
	Modules/EffectEditor/FilePathTypeEditor/FilePathTypeEditor.csproj
	Modules/EffectEditor/LevelTypeEditor/LevelTypeEditor.csproj
	Modules/EffectEditor/SpinEffectEditor/SpinEffectEditor.csproj
	Modules/EffectEditor/TwinkleEffectEditor/TwinkleEffectEditor.csproj
	Modules/Input/DirectXJoystick/DirectXJoystick.csproj
	Modules/Media/Audio/Audio.csproj
	Modules/Output/595/Olsen595.csproj
	Modules/Output/BlinkyLinky/BlinkyLinky.csproj
	Modules/Output/DMXUSBPro/DmxUsbPro.csproj
	Modules/Output/DummyLighting/DummyLighting.csproj
	Modules/Output/E1.31/E131.csproj
	Modules/Output/Hill320/Hill320.csproj
	Modules/Output/PSC/PSC/PSC.csproj
	Modules/Output/Renard/Renard.csproj
	Modules/PostFilter/Color/Color.csproj
	Modules/PostFilter/Grayscale/Grayscale.csproj
	Modules/PreFilter/FadeOut/FadeOut.csproj
	Modules/Property/Grid/Grid.csproj
	Modules/Property/Position/Position.csproj
	Modules/RuntimeBehavior/Recording/Recording.csproj
	Modules/Script/CSharp/CSharp.csproj
	Modules/Script/VB/VB.csproj
	Modules/Sequence/Script/Script.csproj
	Modules/Sequence/Vixen2x/Vixen2x.csproj
	Modules/Timing/Generic/Generic.csproj
	VixenModules.sln
	commonProject.props

+++ Modules/Effect/SetLevel/SetLevelData.cs
@@ -13,7 +13,7 @@ namespace VixenModules.Effect.SetLevel
 	class SetLevelData : ModuleDataModelBase
 	{
 		[DataMember]
-		public double level { get; set; }
+		public float level { get; set; }
 
 		[DataMember]
 		public RGB color { get; set; }
